refactor(mbk): extract pure payer-name matcher to attribution_matcher
PR2's additions pushed attribution_service.py past the file-size growth guard
(was 512 LOC on base; >500 + any growth is blocked). Split the pure,
DB-free matcher (`_levenshtein` + `find_best_match`) into a focused
`attribution_matcher.py` (90 LOC) and import it back. attribution_service.py
is now 492 LOC; behavior is byte-identical (the matcher moved unchanged). The
matcher's existing test file now imports from the new module.
